NETVIRT-1630 migrate to md-sal APIs
handles migration to new mdsal APIs for
    - databroker
    - Listener base classes
    - Readwrite txns

The following changes are the major in this review

1) import org.opendaylight.controller.md.sal.binding.api.DataBroker;
	to
   import org.opendaylight.mdsal.binding.api.DataBroker; (included blueprint)
2) .toJavaUtil(); to be removed
3) import com.google.common.base.Optional
       to
   import java.util.Optional
4) Optional.fromNullable to Optional.ofNullable
5) Optional.absent() to Optional.empty();
6) import
org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
    to
import org.opendaylight.mdsal.common.api.LogicalDatastoreType;
7) import
org.opendaylight.controller.md.sal.binding.api.ReadOnlyTransaction;
    to
  import org.opendaylight.mdsal.binding.api.ReadTransaction;

8) ReadOnlyTransaction readTx = dataBroker.newReadOnlyTransaction()
    to
  ReadTransaction readTx = dataBroker.newReadOnlyTransaction()

9) writeTransaction/rw.submit().get() to writeTransaction.commit().get()

10) ReadFailedException to InterruptedException | ExecutionException

Dependency:
=============
This review is dependent on below reviews

https://git.opendaylight.org/gerrit/c/infrautils/+/89057
https://git.opendaylight.org/gerrit/c/ovsdb/+/87802
https://git.opendaylight.org/gerrit/c/genius/+/89001

+++ elanmanager/impl/src/main/java/org/opendaylight/netvirt/elan/l2gw/recovery/impl/L2GatewayInstanceRecoveryHandler.java
@@ -7,17 +7,16 @@
  */
 package org.opendaylight.netvirt.elan.l2gw.recovery.impl;
 
-import com.google.common.base.Optional;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import javax.inject.Inject;
 import javax.inject.Singleton;
-
-import org.opendaylight.controller.md.sal.binding.api.DataBroker;
-import org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
+import org.opendaylight.genius.datastoreutils.SingleTransactionDataBroker;
 import org.opendaylight.genius.infra.ManagedNewTransactionRunner;
 import org.opendaylight.genius.infra.ManagedNewTransactionRunnerImpl;
-import org.opendaylight.genius.mdsalutil.MDSALUtil;
+import org.opendaylight.mdsal.binding.api.DataBroker;
+import org.opendaylight.mdsal.common.api.LogicalDatastoreType;
 import org.opendaylight.netvirt.elan.l2gw.utils.L2GatewayConnectionUtils;
 import org.opendaylight.serviceutils.srm.ServiceRecoveryInterface;
 import org.opendaylight.serviceutils.srm.ServiceRecoveryRegistry;
@@ -61,30 +60,43 @@ public class L2GatewayInstanceRecoveryHandler implements ServiceRecoveryInterfac
         InstanceIdentifier<L2gateway> l2gatewayInstanceIdentifier = InstanceIdentifier.create(Neutron.class)
                 .child(L2gateways.class).child(L2gateway.class, new L2gatewayKey(uuid));
 
-        Optional<L2gateway> l2gatewayOptional = MDSALUtil.read(dataBroker, LogicalDatastoreType.CONFIGURATION,
-                l2gatewayInstanceIdentifier);
-        L2gateway l2gateway = l2gatewayOptional.get();
+        Optional<L2gateway> l2gatewayOptional = Optional.empty();
+        try {
+            l2gatewayOptional = SingleTransactionDataBroker.syncReadOptional(dataBroker,
+                    LogicalDatastoreType.CONFIGURATION, l2gatewayInstanceIdentifier);
+        } catch (ExecutionException | InterruptedException e) {
+            LOG.error("recoverService: Exception while reading L2gateway DS for the entity {}", entityId, e);
+        }
 
-        List<L2gatewayConnection> l2gatewayConnections = l2GatewayConnectionUtils.getL2GwConnectionsByL2GatewayId(uuid);
-        // Do a delete of l2 gateway connection instances.
-        //No null check required since l2gatewayConnections is known to be non-null.
-        LOG.info("Deleting all l2 gateway connections of l2 gateway instance {}", l2gateway.key());
-        for (L2gatewayConnection l2gatewayConnection: l2gatewayConnections) {
-            InstanceIdentifier<L2gatewayConnection> identifier = InstanceIdentifier.create(Neutron.class)
+        if (l2gatewayOptional.isPresent()) {
+            L2gateway l2gateway = l2gatewayOptional.get();
+
+            List<L2gatewayConnection> l2gatewayConnections = l2GatewayConnectionUtils
+                .getL2GwConnectionsByL2GatewayId(uuid);
+            // Do a delete of l2 gateway connection instances.
+            //No null check required since l2gatewayConnections is known to be non-null.
+            LOG.info("Deleting all l2 gateway connections of l2 gateway instance {}",
+                l2gateway.key());
+            for (L2gatewayConnection l2gatewayConnection : l2gatewayConnections) {
+                InstanceIdentifier<L2gatewayConnection> identifier = InstanceIdentifier
+                    .create(Neutron.class)
                     .child(L2gatewayConnections.class)
                     .child(L2gatewayConnection.class, l2gatewayConnection.key());
-            try {
-                LOG.info("Deleting l2 gateway connection {}",l2gatewayConnection.key());
-                managedNewTransactionRunner.callWithNewWriteOnlyTransactionAndSubmit(
-                    tx -> tx.delete(LogicalDatastoreType.CONFIGURATION, identifier)).get();
-                LOG.info("Recreating l2 gateway connection {}",l2gatewayConnection.key());
-                managedNewTransactionRunner.callWithNewWriteOnlyTransactionAndSubmit(
-                    tx -> tx.put(LogicalDatastoreType.CONFIGURATION, identifier, l2gatewayConnection)).get();
-            } catch (InterruptedException | ExecutionException e) {
-                LOG.error("Service recovery failed for l2gw {}", entityId);
+                try {
+                    LOG.info("Deleting l2 gateway connection {}", l2gatewayConnection.key());
+                    managedNewTransactionRunner.callWithNewWriteOnlyTransactionAndSubmit(
+                        tx -> tx.delete(LogicalDatastoreType.CONFIGURATION, identifier)).get();
+                    LOG.info("Recreating l2 gateway connection {}", l2gatewayConnection.key());
+                    managedNewTransactionRunner.callWithNewWriteOnlyTransactionAndSubmit(
+                        tx -> tx.put(LogicalDatastoreType.CONFIGURATION, identifier,
+                            l2gatewayConnection)).get();
+                } catch (InterruptedException | ExecutionException e) {
+                    LOG.error("Service recovery failed for l2gw {}", entityId);
+                }
             }
+            LOG.info("Finished recreation of all l2 gateway connections of l2 gateway instance {}",
+                l2gateway.key());
         }
-        LOG.info("Finished recreation of all l2 gateway connections of l2 gateway instance {}", l2gateway.key());
     }
 
     public String buildServiceRegistryKey() {
